Remove dependencies[1] incompatible with the current Swift 2.2 beta.
[1]:  Summary:

These dependencies were incompatible with Swift 2.2 beta:

Nimble 3.0.0
FBSnapshotTestCase
Nimble-Snapshots

The stubbed-out placeholder tests in Tests.swift depended on these, so I removed that code, too.

Quick 0.8.0 didn't have any problem with Swift 2.2, but since the removed test code was the only thing that used it, I eliminated that dependency also.

So the only dependency is AlamoFire, which I am keeping for now.